Early morning study session of day 12/1
Today was a very good day. I learned about creating, deleting, and changing an archive in Python. After that, I learned about JSON, and it was pretty cool. I even saw some examples of javascript. It was pretty hard, but it was cool. I will try later on. This was, till now, the best and most productive night of study.
I started studying at 11:20 p.m. and went till 4:40 a.m., resulting in 5 hours and 20 minutes of studying.