First Commit via CLI test
I was not able to study too much today because I did an exercise that took a lot of time to do. Why did it take so long? I tried to do it by myself, I think thats why, but it was good because I did it; I managed to complete it without any help, and it was very gratifying to do so. I tried to learn about commits on Github via CLI; actually, this is a commit via CLI. I did around 4 hours of study, and this week I will finish the fourth module. I am really excited for the next one