Early morning study session on 12/07/2023
Today I spent the night chatting with BRK. I didnt study much, but the talk was great and gave me a lot of perspective about my future in programming and in life too. About coding, I added a implemment to the list of tasks. At the end of adding tasks to the list, you can finish, and then it will create a JSON with the list of tasks so you can keep track of what you need to do.