Late-night study session from December 12th to December 13th.
I started studying today at 12:00 p.m. I did not have a lot of time to study, but I tried my best. The classes are really challenging; this module will demand a lot of effort and require rewatching some classes several times, but it might be worth it; let's see. I completed an exercise that the teacher sent us, and in the first part, I somehow managed to do it exactly the same way he did; I just used different names for the variables and classes, but the logic was the same. I felt really proud of myself for that; it made me feel like I was making progress. In the second part of the exercise, I had some idea of what to do, but I was not able to complete it. So, I watched the resolution of the exercise, and in the middle of the resolution, I got an idea of how to do it. I was then able to complete it on my own, but this time, it was different from the teacher's approach, in terms of logic, I mean, but it was still great. Tomorrow, I will start studying earlier to have more time for it.